+++ src/app.ts
@@ -4,6 +4,7 @@ import { httpLogger } from "./logger/index.js";
 import routes from "./routes/index.js";
 import { errorMiddleware } from "./middlewares/error.middleware.js";
 import { globalRateLimiter } from "./middlewares/rateLimit.middleware.js";
+import { configureSwagger } from "./docs/swagger.js";
 
 const app = express();
 
@@ -13,6 +14,7 @@ app.use(express.json());
 app.use(httpLogger);
 app.use(globalRateLimiter);
 app.use("/api/v1", routes);
+configureSwagger(app);
 app.use(errorMiddleware);
 
 export default app;