+++ src/docs/tags.ts
@@ -1,0 +1,16 @@
+export const tags = [
+  {
+    name: "Authentication",
+    description: "Authentication and profile management.",
+  },
+
+  {
+    name: "Categories",
+    description: "Manage transaction categories.",
+  },
+
+  {
+    name: "Transactions",
+    description: "Manage income and expense transactions.",
+  },
+];